assign2/requirements.txt
torch
torchvision
ultralytics
opencv-python
numpy
pandas
Pillow
streamlit
